+++ circuits/circom2gnark.go
@@ -30,3 +30,14 @@ func Circom2GnarkProof(vkey []byte, circomProof, pubSignals string) (*parser.Gna
 	}
 	return proof, nil
 }
+
+// Circom2GnarkPlaceholder function is a wrapper to convert the circom ballot
+// circuit to a gnark recursion placeholder, it returns the resulting
+// placeholders for the recursion.
+func Circom2GnarkPlaceholder(vkey []byte) (*parser.GnarkRecursionPlaceholders, error) {
+	gnarkVKeyData, err := parser.UnmarshalCircomVerificationKeyJSON(vkey)
+	if err != nil {
+		return nil, err
+	}
+	return parser.PlaceholdersForRecursion(gnarkVKeyData, BallotProofNPubInputs, true)
+}
